Columbus State CC Capstone Project
I worked on a team with two other classmates in order to develop a
mock-up of an ERP system. My colleagues were responsible for developing
organizational documentation, databases, and the like, while I was
tasked with developing the GUI and applications for the org, as well as
training documents for end user utilization of the programs.